x86: Add __sigsetjmp_cancel and __setjmp_cancel
Functions, like LIBC_START_MAIN, START_THREAD_DEFN as well as these
with thread cancellation, call setjmp, but never return to their
callers after longjmp returns.  This patch adds <bits/setjmp-cancel.h>
and <setjmp-cancelP.h> to provide a version of setjmp family functions,
__setjmp_cancel and __sigsetjmp_cancel, which are used to implement
thread cancellation.  The default __setjmp_cancel and __sigsetjmp_cancel
are defined as setjmp and __sigsetjmp, respectively, which are the same
as before.

On x86, __sigsetjmp_cancel, which is a public function used by macros
in <pthread.h>, and __setjmp_cancel, which is a private interface for
cancellation implementation in libpthread, are added to avoid saving
and restoring shadow stack register.  A new version of __sigsetjmp is
added to save and restore shadow stack register. The compatible version
of __sigsetjmp, which doesn't save and restore shadow stack register,
is provided for shared libraries built with glibc 2.27 or older.
__libc_longjmp, which is a private interface for cancellation
implementation in libpthread, is changed to call __longjmp_cancel
instead of __longjmp.

NB: Shared libraries built with glibc 2.27 or older are compatible with
glibc 2.28.  Relocatable objects compiled against glibc 2.27 or older
are incompatible with glibc 2.28 since they reference the older version
of __sigsetjmp.

Tested with build-many-glibcs.py.

	* bits/setjmp-cancel.h: New file.
	* sysdeps/generic/setjmp-cancelP.h: Likewise.
	* sysdeps/x86/__longjmp_cancel.S: Likewise.
	* sysdeps/x86/bits/setjmp-cancel.h: Likewise.
	* sysdeps/x86/bsd-_setjmp_cancel.S: Likewise.
	* sysdeps/x86/longjmp.c: Likewise.
	* sysdeps/x86/setjmp-cancelP.h: Likewise.
	* sysdeps/x86/setjmp_cancel.S: Likewise.
	* sysdeps/x86_64/setjmp_cancel.S: Likewise.
	* csu/libc-start.c (LIBC_START_MAIN): Replace setjmp with
	__setjmp_cancel.
	* nptl/pthread_create.c (START_THREAD_DEFN): Likewise.
	* include/setjmp.h: Include <setjmp-cancelP.h> and
	<bits/setjmp-cancel.h>.
	* nptl/Makefile (headers): Add bits/setjmp-cancel.h.
	* setjmp/longjmp.c (__libc_longjmp): Don't define alias if
	defined.
	* sysdeps/i386/bsd-_setjmp.S (_setjmp): Don't add hidden_def.
	* sysdeps/x86_64/bsd-_setjmp.S (_setjmp): Likewise.
	* sysdeps/i386/setjmp.S (__sigsetjmp): Don't add hidden_def and
	redefined to __redirect___sigsetjmp in libc.so if __sigsetjmp
	is undefined.
	(__GI___sigsetjmp): New.  Defined if __redirect___sigsetjmp is
	defined.
	(__sigsetjmp): Add to GLIBC_2_28 if __redirect___sigsetjmp is
	defined.
	* sysdeps/x86_64/setjmp.S: Likewise.
	* sysdeps/nptl/pthread.h (pthread_cleanup_push): Replace
	__sigsetjmp with __sigsetjmp_cancel.
	(pthread_cleanup_push_defer_np): Likewise.
	(__jmp_buf_tag): Removed.
	(__sigsetjmp): Likewise.
	Include <bits/setjmp-cancel.h>.
	* sysdeps/unix/sysv/linux/i386/libc.abilist: Regenerated.
	* sysdeps/unix/sysv/linux/x86_64/64/libc.abilist: Likewise.
	* sysdeps/unix/sysv/linux/x86_64/x32/libc.abilist: Likewise.
	* sysdeps/x86/Makefile (sysdep_routines): Add
	bsd-_setjmp_cancel, __longjmp_cancel and setjmp_cancel.
	* sysdeps/x86/Versions (__sigsetjmp): Add to GLIBC_2.28
	in libc.
	(__sigsetjmp_cancel): Likewise.
	(__setjmp_cancel): Add to GLIBC_PRIVATE in libc.
	* sysdeps/x86_64/bsd-_setjmp.S (SIGSETJMP): New.
	(_setjmp): Jmp to SIGSETJMP.