Initial empty CPU run-time library for C: libcpu-rt-c
Add the --enable-cpu-rt configure option to enable the CPU run-time
library for C, libcpu-rt-c.  libcpu-rt-c contains a subset of the C
library with the optimized functions.  The resulting libcpu-rt-c.so
should be binary compatible with older versions of libc.so so that
libcpu-rt-c.so can be used with LD_PRELOAD or linked directly with
applications.

	* Makeconfig (rpath-dirs): Add cpu-rt-c.
	(all-subdirs): Likewise.
	* configure: Regenerated.
	* configure.ac: Add --enable-cpu-rt.
	(build-cpu-rt): New LIBC_CONFIG_VAR.
	* shlib-versions: Add libcpu_rt-c.
	* cpu-rt-c/Makefile: New file.